Extract server trace route helpers
